labview: Timed loop fix
Timed loops were erroring out because cgroups were not enabled (and mounted).
Update liblvalarms to ignore the error returned from the cgroup functions so
that timed loops can initialize properly.

Note that if the proper cgroups are not present the CPU pool VIs will not
work, but CPU affinity will still function properly.