core-image-minimal-chroot: add new recipe
This recipe is meant to create a rootfs sufficient to be used as a chroot on
another host OS. The key was basing the recipe off of core-image-minimal and
using the linux-dummy recipe since no kernel or modules are needed in a
chroot. Bash was also added as a dependency since interaction with the chroot
will often be via command line.