Fix dependency warning when bitbaking lv-appweb-support.
Workaround bitbake's overzealous caching of built images;
it wasn't noticing when files in the pre-built dir hierarchy
were changing because they weren't explicitly references in
the SRC_UI list.  We now add a manifest file with checksums,
added to SRC_URI but not actually installed, so it will now notice
if a file changes and really rebuild.